RAB-1236: Create CheckStorage component & refactor tests
