+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageConfigurator/AmazonS3StorageConfigurator.test.tsx
@@ -1,26 +1,24 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
-import {screen, act} from '@testing-library/react';
+import {screen} from '@testing-library/react';
 import {renderWithProviders, ValidationError} from '@akeneo-pim-community/shared';
-import {AmazonS3Storage, LocalStorage} from '../model';
 import {AmazonS3StorageConfigurator} from './AmazonS3StorageConfigurator';
+import {AmazonS3Storage, LocalStorage} from '../../models';
 
-beforeEach(() => {
-  global.fetch = mockFetch;
-});
+jest.mock('./CheckStorageConnection', () => ({
+  CheckStorageConnection: () => <button>Check connection</button>,
+}));
 
-const mockFetch = jest.fn().mockImplementation(async (route: string) => {
-  switch (route) {
-    case 'pimee_job_automation_get_storage_connection_check':
-      return {
-        ok: true,
-      };
-    default:
-      throw new Error();
-  }
-});
+const storage: AmazonS3Storage = {
+  type: 'amazon_s3',
+  file_path: '/tmp/file.xlsx',
+  region: 'eu-west-1',
+  bucket: '',
+  key: 'a_key',
+  secret: 'my_s3cr3t',
+};
 
-test('it renders the amazon s3 storage configurator', async () => {
+test('it renders the amazon s3 storage configurator', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/file.xlsx',
@@ -30,24 +28,22 @@ test('it renders the amazon s3 storage configurator', async () => {
     secret: 'my_s3cr3t',
   };
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   expect(screen.getByDisplayValue('/tmp/file.xlsx')).toBeInTheDocument();
   expect(screen.getByDisplayValue('eu-west-1')).toBeInTheDocument();
   expect(screen.getByDisplayValue('my_bucket')).toBeInTheDocument();
 });
 
-test('it allows user to fill file_path field', async () => {
+test('it allows user to fill file_path field', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/test.xls',
@@ -59,17 +55,15 @@ test('it allows user to fill file_path field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const file_pathInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.file_path.label pim_common.required_label'
@@ -79,7 +73,7 @@ test('it allows user to fill file_path field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, file_path: '/tmp/test.xlsx'});
 });
 
-test('it allows user to fill region field', async () => {
+test('it allows user to fill region field', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/file.xlsx',
@@ -91,17 +85,15 @@ test('it allows user to fill region field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const regionInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.region.label pim_common.required_label'
@@ -111,29 +103,18 @@ test('it allows user to fill region field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, region: 'eu-west-1'});
 });
 
-test('it allows user to fill bucket field', async () => {
-  const storage: AmazonS3Storage = {
-    type: 'amazon_s3',
-    file_path: '/tmp/file.xlsx',
-    region: 'eu-west-1',
-    bucket: '',
-    key: 'a_key',
-    secret: 'my_s3cr3t',
-  };
-
+test('it allows user to fill bucket field', () => {
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.paste(
     screen.getByLabelText('pim_import_export.form.job_instance.storage_form.bucket.label pim_common.required_label'),
@@ -146,7 +127,7 @@ test('it allows user to fill bucket field', async () => {
   });
 });
 
-test('it allows user to fill key field', async () => {
+test('it allows user to fill key field', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/file.xlsx',
@@ -158,17 +139,15 @@ test('it allows user to fill key field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const keyInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.key.label pim_common.required_label'
@@ -178,7 +157,7 @@ test('it allows user to fill key field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, key: 'a_key'});
 });
 
-test('it allows user to fill secret field', async () => {
+test('it allows user to fill secret field', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/file.xlsx',
@@ -190,17 +169,15 @@ test('it allows user to fill secret field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const secretInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.secret.label pim_common.required_label'
@@ -219,17 +196,15 @@ test('it hides secret field if the secret is obfuscated', () => {
     key: 'a_key',
   };
 
-  act(() => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   const secretInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.secret.label pim_common.required_label'
@@ -249,23 +224,23 @@ test('it can edit the secret field if the secret is obfuscated', () => {
   };
 
   const onStorageChange = jest.fn();
-  act(() => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.click(screen.getByText('pim_common.edit'));
+
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, secret: ''});
 });
 
-test('it throws an exception when passing a non amazon s3 storage', async () => {
+test('it throws an exception when passing a non amazon s3 storage', () => {
   const mockedConsole = jest.spyOn(console, 'error').mockImplementation();
 
   const storage: LocalStorage = {
@@ -288,7 +263,7 @@ test('it throws an exception when passing a non amazon s3 storage', async () => 
   mockedConsole.mockRestore();
 });
 
-test('it displays validation errors', async () => {
+test('it displays validation errors', () => {
   const storage: AmazonS3Storage = {
     type: 'amazon_s3',
     file_path: '/tmp/file.xlsx',
@@ -336,17 +311,15 @@ test('it displays validation errors', async () => {
     },
   ];
 
-  await act(async () => {
-    renderWithProviders(
-      <AmazonS3StorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={validationErrors}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <AmazonS3StorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={validationErrors}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   expect(screen.getByText('error.key.a_file_path_error')).toBeInTheDocument();
   expect(screen.getByText('error.key.a_region_error')).toBeInTheDocument();
@@ -355,74 +328,16 @@ test('it displays validation errors', async () => {
   expect(screen.getByText('error.key.a_secret_error')).toBeInTheDocument();
 });
 
-test('it can check connection', async () => {
-  const storage: AmazonS3Storage = {
-    type: 'amazon_s3',
-    file_path: '/tmp/file.xlsx',
-    region: 'eu-west-1',
-    bucket: '',
-    key: 'a_key',
-    secret: 'my_s3cr3t',
-  };
-
-  const onStorageChange = jest.fn();
-
+test('it can check connection', () => {
   renderWithProviders(
     <AmazonS3StorageConfigurator
       jobInstanceCode="csv_product_export"
       storage={storage}
       fileExtension="xlsx"
       validationErrors={[]}
-      onStorageChange={onStorageChange}
+      onStorageChange={jest.fn()}
     />
   );
 
-  const checkButton = screen.getByText('pim_import_export.form.job_instance.connection_checker.label');
-  await act(async () => {
-    userEvent.click(checkButton);
-  });
-
-  expect(checkButton).toBeDisabled();
-});
-
-test('it can check connection, display message if error', async () => {
-  mockFetch.mockImplementation((route: string) => {
-    switch (route) {
-      case 'pimee_job_automation_get_storage_connection_check':
-        return {
-          ok: false,
-        };
-      default:
-        throw new Error();
-    }
-  });
-
-  const storage: AmazonS3Storage = {
-    type: 'amazon_s3',
-    file_path: '/tmp/file.xlsx',
-    region: 'eu-west-1',
-    bucket: 'my_bucket',
-    key: 'a_key',
-    secret: 'my_s3cr3t',
-  };
-
-  const onStorageChange = jest.fn();
-
-  renderWithProviders(
-    <AmazonS3StorageConfigurator
-      jobInstanceCode="csv_product_export"
-      storage={storage}
-      fileExtension="xlsx"
-      validationErrors={[]}
-      onStorageChange={onStorageChange}
-    />
-  );
-
-  const checkButton = screen.getByText('pim_import_export.form.job_instance.connection_checker.label');
-  await act(async () => {
-    userEvent.click(checkButton);
-  });
-
-  expect(checkButton).not.toBeDisabled();
-  expect(screen.getByText('pim_import_export.form.job_instance.connection_checker.exception')).toBeInTheDocument();
+  expect(screen.getByText('Check connection')).toBeInTheDocument();
 });
