+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageConfigurator/SftpStorageConfigurator.test.tsx
@@ -1,31 +1,19 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
-import {screen, act} from '@testing-library/react';
+import {screen} from '@testing-library/react';
 import {renderWithProviders, ValidationError} from '@akeneo-pim-community/shared';
-import {LocalStorage, SftpStorage} from '../model';
 import {SftpStorageConfigurator} from './SftpStorageConfigurator';
+import {SftpStorage, LocalStorage} from '../../models';
 
-beforeEach(() => {
-  global.fetch = mockFetch;
-});
+jest.mock('./CheckStorageConnection', () => ({
+  CheckStorageConnection: () => <button>Check connection</button>,
+}));
 
-const mockFetch = jest.fn().mockImplementation(async (route: string) => {
-  switch (route) {
-    case 'pimee_job_automation_get_public_key':
-      return {
-        ok: true,
-        json: async () => '-----BEGIN CERTIFICATE-----publickey-----END CERTIFICATE-----',
-      };
-    case 'pimee_job_automation_get_storage_connection_check':
-      return {
-        ok: true,
-      };
-    default:
-      throw new Error();
-  }
-});
+jest.mock('../../hooks/useGetPublicKey', () => ({
+  useGetPublicKey: () => '-----BEGIN CERTIFICATE-----publickey-----END CERTIFICATE-----',
+}));
 
-test('it renders the sftp storage configurator', async () => {
+test('it renders the sftp storage configurator', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '/tmp/file.xlsx',
@@ -37,23 +25,21 @@ test('it renders the sftp storage configurator', async () => {
     password: 'root',
   };
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   expect(screen.getByDisplayValue('/tmp/file.xlsx')).toBeInTheDocument();
   expect(screen.getByDisplayValue('c1:91:5e:42:55:5c:74:65:b6:12:32:7e:1f:6d:80:3e')).toBeInTheDocument();
 });
 
-test('it allows user to fill file_path field', async () => {
+test('it allows user to fill file_path field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '/tmp/test.xls',
@@ -66,17 +52,15 @@ test('it allows user to fill file_path field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const file_pathInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.file_path.label pim_common.required_label'
@@ -86,7 +70,7 @@ test('it allows user to fill file_path field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, file_path: '/tmp/test.xlsx'});
 });
 
-test('it allows user to fill host field', async () => {
+test('it allows user to fill host field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -99,17 +83,15 @@ test('it allows user to fill host field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const hostInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.host.label pim_common.required_label'
@@ -119,7 +101,7 @@ test('it allows user to fill host field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, host: 'example.com'});
 });
 
-test('it allows user to fill fingerprint field', async () => {
+test('it allows user to fill fingerprint field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -132,17 +114,15 @@ test('it allows user to fill fingerprint field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.paste(
     screen.getByLabelText('pim_import_export.form.job_instance.storage_form.fingerprint.label'),
@@ -155,7 +135,7 @@ test('it allows user to fill fingerprint field', async () => {
   });
 });
 
-test('it removes fingerprint from model when clearing input', async () => {
+test('it removes fingerprint from model when clearing input', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -169,17 +149,15 @@ test('it removes fingerprint from model when clearing input', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.clear(screen.getByLabelText('pim_import_export.form.job_instance.storage_form.fingerprint.label'));
 
@@ -189,7 +167,7 @@ test('it removes fingerprint from model when clearing input', async () => {
   });
 });
 
-test('it allows user to fill port field', async () => {
+test('it allows user to fill port field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -202,17 +180,15 @@ test('it allows user to fill port field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const portInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.port.label pim_common.required_label'
@@ -222,7 +198,7 @@ test('it allows user to fill port field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, port: 22});
 });
 
-test('it allows user to change login type', async () => {
+test('it allows user to change login type', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -235,17 +211,15 @@ test('it allows user to change login type', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.click(screen.getByLabelText('pim_import_export.form.job_instance.storage_form.login_type.label'));
   userEvent.click(screen.getByText('pim_import_export.form.job_instance.storage_form.login_type.private_key'));
@@ -256,7 +230,7 @@ test('it allows user to change login type', async () => {
   });
 });
 
-test('it displays a public key field', async () => {
+test('it displays a public key field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -269,24 +243,22 @@ test('it displays a public key field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const publicKeyField = screen.getByTestId('publicKey');
   expect(publicKeyField).toBeInTheDocument();
   expect(publicKeyField).toHaveDisplayValue('-----BEGIN CERTIFICATE-----publickey-----END CERTIFICATE-----');
 });
 
-test('it copy to clipboard a public key', async () => {
+test('it copy to clipboard a public key', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -306,24 +278,23 @@ test('it copy to clipboard a public key', async () => {
   });
   jest.spyOn(navigator.clipboard, 'writeText');
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
+
   userEvent.click(screen.getByTestId('copyToClipboard'));
   expect(navigator.clipboard.writeText).toHaveBeenCalledWith(
     '-----BEGIN CERTIFICATE-----publickey-----END CERTIFICATE-----'
   );
 });
 
-test('it allows user to fill username field', async () => {
+test('it allows user to fill username field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -336,17 +307,15 @@ test('it allows user to fill username field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const usernameInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.username.label pim_common.required_label'
@@ -356,7 +325,7 @@ test('it allows user to fill username field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, username: 'root'});
 });
 
-test('it allows user to fill password field', async () => {
+test('it allows user to fill password field', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -369,17 +338,15 @@ test('it allows user to fill password field', async () => {
 
   const onStorageChange = jest.fn();
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   const passwordInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.password.label pim_common.required_label'
@@ -389,7 +356,7 @@ test('it allows user to fill password field', async () => {
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, password: 'root'});
 });
 
-test('it hides password field if the password is obfuscated', async () => {
+test('it hides password field if the password is obfuscated', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -399,17 +366,15 @@ test('it hides password field if the password is obfuscated', async () => {
     username: '',
   };
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   const connectionStringInput = screen.getByLabelText(
     'pim_import_export.form.job_instance.storage_form.password.label pim_common.required_label'
@@ -419,7 +384,7 @@ test('it hides password field if the password is obfuscated', async () => {
   expect(connectionStringInput).toHaveValue('••••••••');
 });
 
-test('it can edit the password field if the password is obfuscated', async () => {
+test('it can edit the password field if the password is obfuscated', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '',
@@ -430,23 +395,22 @@ test('it can edit the password field if the password is obfuscated', async () =>
   };
 
   const onStorageChange = jest.fn();
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={[]}
-        onStorageChange={onStorageChange}
-      />
-    );
-  });
+
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={[]}
+      onStorageChange={onStorageChange}
+    />
+  );
 
   userEvent.click(screen.getByText('pim_common.edit'));
   expect(onStorageChange).toHaveBeenLastCalledWith({...storage, password: ''});
 });
 
-test('it throws an exception when passing a non-sftp storage', async () => {
+test('it throws an exception when passing a non-sftp storage', () => {
   const mockedConsole = jest.spyOn(console, 'error').mockImplementation();
 
   const storage: LocalStorage = {
@@ -469,7 +433,7 @@ test('it throws an exception when passing a non-sftp storage', async () => {
   mockedConsole.mockRestore();
 });
 
-test('it displays validation errors', async () => {
+test('it displays validation errors', () => {
   const storage: SftpStorage = {
     type: 'sftp',
     file_path: '/tmp/file.xlsx',
@@ -526,17 +490,15 @@ test('it displays validation errors', async () => {
     },
   ];
 
-  await act(async () => {
-    renderWithProviders(
-      <SftpStorageConfigurator
-        jobInstanceCode="csv_product_export"
-        storage={storage}
-        fileExtension="xlsx"
-        validationErrors={validationErrors}
-        onStorageChange={jest.fn()}
-      />
-    );
-  });
+  renderWithProviders(
+    <SftpStorageConfigurator
+      jobInstanceCode="csv_product_export"
+      storage={storage}
+      fileExtension="xlsx"
+      validationErrors={validationErrors}
+      onStorageChange={jest.fn()}
+    />
+  );
 
   expect(screen.getByText('error.key.a_file_path_error')).toBeInTheDocument();
   expect(screen.getByText('error.key.a_host_error')).toBeInTheDocument();
@@ -544,83 +506,4 @@ test('it displays validation errors', async () => {
   expect(screen.getByText('error.key.a_port_error')).toBeInTheDocument();
   expect(screen.getByText('error.key.an_username_error')).toBeInTheDocument();
   expect(screen.getByText('error.key.a_password_error')).toBeInTheDocument();
-});
-
-test('it can check connection', async () => {
-  const storage: SftpStorage = {
-    type: 'sftp',
-    file_path: '/tmp/file.xlsx',
-    host: 'example.com',
-    port: 22,
-    login_type: 'password',
-    username: 'root',
-    password: 'root',
-  };
-
-  const onStorageChange = jest.fn();
-
-  renderWithProviders(
-    <SftpStorageConfigurator
-      jobInstanceCode="csv_product_export"
-      storage={storage}
-      fileExtension="xlsx"
-      validationErrors={[]}
-      onStorageChange={onStorageChange}
-    />
-  );
-
-  const checkButton = screen.getByText('pim_import_export.form.job_instance.connection_checker.label');
-  await act(async () => {
-    userEvent.click(checkButton);
-  });
-
-  expect(checkButton).toBeDisabled();
-});
-
-test('it can check connection, display message if error', async () => {
-  mockFetch.mockImplementation((route: string) => {
-    switch (route) {
-      case 'pimee_job_automation_get_public_key':
-        return {
-          ok: true,
-          json: async () => '-----BEGIN CERTIFICATE-----publickey-----END CERTIFICATE-----',
-        };
-      case 'pimee_job_automation_get_storage_connection_check':
-        return {
-          ok: false,
-        };
-      default:
-        throw new Error();
-    }
-  });
-
-  const storage: SftpStorage = {
-    type: 'sftp',
-    file_path: '/tmp/file.xlsx',
-    host: 'example.com',
-    port: 22,
-    login_type: 'password',
-    username: 'root',
-    password: 'root',
-  };
-
-  const onStorageChange = jest.fn();
-
-  renderWithProviders(
-    <SftpStorageConfigurator
-      jobInstanceCode="csv_product_export"
-      storage={storage}
-      fileExtension="xlsx"
-      validationErrors={[]}
-      onStorageChange={onStorageChange}
-    />
-  );
-
-  const checkButton = screen.getByText('pim_import_export.form.job_instance.connection_checker.label');
-  await act(async () => {
-    userEvent.click(checkButton);
-  });
-
-  expect(checkButton).not.toBeDisabled();
-  expect(screen.getByText('pim_import_export.form.job_instance.connection_checker.exception')).toBeInTheDocument();
 });
