+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageForm.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import {screen, act} from '@testing-library/react';
 import {renderWithProviders, ValidationError, useFeatureFlags} from '@akeneo-pim-community/shared';
 import {StorageForm} from './StorageForm';
-import {LocalStorage, NoneStorage, SftpStorage} from './model';
+import {NoneStorage, LocalStorage, SftpStorage} from '../models';
 
 const mockedUseFeatureFlags = useFeatureFlags as jest.Mock;
