+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/index.ts
@@ -1,2 +1,1 @@
 export * from './StorageForm';
-export * from './model';
