+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/models/Storage.test.ts
@@ -6,7 +6,60 @@ import {
   getDefaultFilePath,
   localStorageIsEnabled,
   additionalStorageIsEnabled,
-} from './model';
+  isAmazonS3Storage,
+  isGoogleCloudStorage,
+  isLocalStorage,
+  isMicrosoftAzureStorage,
+  isSftpStorage,
+  AmazonS3Storage,
+  GoogleCloudStorage,
+  LocalStorage,
+  MicrosoftAzureStorage,
+  SftpStorage,
+  isSftpConnectionFieldFulfilled,
+  isGoogleCloudConnectionFieldFulfilled,
+  isAmazonS3ConnectionFieldFulfilled,
+  isMicrosoftAzureConnectionFieldFulfilled,
+} from './Storage';
+
+const localStorage: LocalStorage = {
+  type: 'local',
+  file_path: '/tmp/test.xlsx',
+};
+
+const sftpStorage: SftpStorage = {
+  type: 'sftp',
+  host: 'example.com',
+  port: 22,
+  login_type: 'password',
+  username: 'test',
+  password: 'test',
+  file_path: '/tmp/test.xlsx',
+};
+
+const amazonS3Storage: AmazonS3Storage = {
+  type: 'amazon_s3',
+  region: 'eu-west-3',
+  bucket: 'a_bucket',
+  key: 'test',
+  secret: 'test',
+  file_path: '/tmp/test.xlsx',
+};
+
+const microsoftAzureStorage: MicrosoftAzureStorage = {
+  type: 'microsoft_azure',
+  connection_string: 'agagag',
+  container_name: 'ahaha',
+  file_path: '/tmp/test.xlsx',
+};
+
+const googleCloudStorage: GoogleCloudStorage = {
+  type: 'google_cloud_storage',
+  file_path: '/tmp/test.xlsx',
+  project_id: 'eu-west-3',
+  service_account: '{"type": "service_account"}',
+  bucket: 'a_bucket',
+};
 
 const featureFlagCollection = {
   import_export_local_storage: false,
@@ -118,4 +171,51 @@ test('it check if additionnal storage is enabled', () => {
   expect(additionalStorageIsEnabled(featureFlags)).toBe(false);
   enableFeatureFlag('import_export_additional_storage');
   expect(additionalStorageIsEnabled(featureFlags)).toBe(true);
+});
+
+test('it says if a storage is a local storage', () => {
+  expect(isLocalStorage(localStorage)).toBe(true);
+  expect(isLocalStorage(sftpStorage)).toBe(false);
+  expect(isLocalStorage(amazonS3Storage)).toBe(false);
+});
+
+test('it says if a storage is a sftp storage', () => {
+  expect(isSftpStorage(sftpStorage)).toBe(true);
+  expect(isSftpStorage({...sftpStorage, fingerprint: 'c1:91:5e:42:55:5c:74:65:b6:12:32:7e:1f:6d:80:3e'})).toBe(true);
+  expect(isSftpStorage(amazonS3Storage)).toBe(false);
+  expect(isSftpStorage(localStorage)).toBe(false);
+});
+
+test('it says if a storage is an amazon s3 storage', () => {
+  expect(isAmazonS3Storage(amazonS3Storage)).toBe(true);
+  expect(isAmazonS3Storage(sftpStorage)).toBe(false);
+  expect(isSftpStorage(localStorage)).toBe(false);
+});
+
+test('it says if a storage is a microsoft azure storage', () => {
+  expect(isMicrosoftAzureStorage(microsoftAzureStorage)).toBe(true);
+  expect(isMicrosoftAzureStorage(sftpStorage)).toBe(false);
+  expect(isSftpStorage(localStorage)).toBe(false);
+});
+
+test('it says if a storage is a google cloud storage', () => {
+  expect(isGoogleCloudStorage(googleCloudStorage)).toBe(true);
+  expect(isGoogleCloudStorage(sftpStorage)).toBe(false);
+  expect(isSftpStorage(localStorage)).toBe(false);
+});
+
+test('it can check that a SFTP storage is filled', () => {
+  expect(isSftpConnectionFieldFulfilled(sftpStorage)).toBe(true);
+});
+
+test('it can check that a GCS storage is filled', () => {
+  expect(isGoogleCloudConnectionFieldFulfilled(googleCloudStorage)).toBe(true);
+});
+
+test('it can check that a Amazon S3 storage is filled', () => {
+  expect(isAmazonS3ConnectionFieldFulfilled(amazonS3Storage)).toBe(true);
+});
+
+test('it can check that a Azure storage is filled', () => {
+  expect(isMicrosoftAzureConnectionFieldFulfilled(microsoftAzureStorage)).toBe(true);
 });
