+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/models/index.ts
@@ -1,0 +1,1 @@
+export * from './Storage';
